Added exceptions in ContextService
If you try to use ContextService::initializeShopContext() now but the container has no service called `shop` a RuntimeException is thrown. Formerly when you tried this a fatal error would occur because `$this->container->get('shop')` returns null and then some method would be called on null. This PR makes it possible for external callers to react on such an error.